requirement.txt
openai
yt_dlp
python-dotenv
flask
flask_cors
moviepy
pydub
asyncio